Style of remaining columns with no values
When setting a style for all columns the reader only reads this style for filled columns. The style of the remaining columns is dropped.
Example: You activate border lines for all cells, but you fill only the first three columns. After reading and writing it with ABAP2XLSX, only the first three columns have border lines.

This is fixed by setting the default style here.